Wire htslib faidx into CLI for real FASTA reference loading
- Link htslib in build.zig (include + library paths to ../htslib)
- Export vcf_htslib module from lib.zig
- Add htsFaidxFetchAdapter bridging FetchSeqFn to HtsFaidx
- Open faidx in main.zig and pass to CsqContext
- Fix faidx_fetch_seq64 len type (i64 not c_int)

The binary now loads FASTA references and runs on real test data.
BCSQ consequence annotation pipeline executes but output injection
needs debugging for coordinate mapping between GFF and VCF.

+++ zig/src/vcf/htslib.zig
@@ -137,7 +137,7 @@ pub const HtsFaidx = struct {
 
     /// Fetch a sequence region. Caller owns the returned slice.
     pub fn fetchSeq(self: *const HtsFaidx, alloc: std.mem.Allocator, seq: [*:0]const u8, beg: i64, end: i64) ![]u8 {
-        var len: c_int = 0;
+        var len: i64 = 0;
         const raw = c.faidx_fetch_seq64(self.fai, seq, beg, end, &len);
         if (raw == null) return error.FaidxFetchFailed;
         defer std.c.free(raw);
